+++ Sources/ParseSwift/API/API+Command.swift
@@ -430,17 +430,17 @@ internal extension API.Command {
         let mapper = { (mapperData: Data) -> T in
             var updatedObject = object
             updatedObject.originalData = nil
-            let object = try ParseCoding
+            updatedObject = try ParseCoding
                 .jsonDecoder()
                 .decode(ReplaceResponse.self, from: mapperData)
                 .apply(to: updatedObject)
             guard let originalData = data,
                   let original = try? ParseCoding.jsonDecoder().decode(T.self,
                                                                        from: originalData),
-                  original.hasSameObjectId(as: object) else {
-                      return object
+                  original.hasSameObjectId(as: updatedObject) else {
+                      return updatedObject
                   }
-            return try object.merge(with: original)
+            return try updatedObject.merge(with: original)
         }
         return API.Command<T, T>(method: .PUT,
                                  path: object.endpoint,
@@ -456,17 +456,17 @@ internal extension API.Command {
         let mapper = { (mapperData: Data) -> T in
             var updatedObject = object
             updatedObject.originalData = nil
-            let object = try ParseCoding
+            updatedObject = try ParseCoding
                 .jsonDecoder()
                 .decode(UpdateResponse.self, from: mapperData)
                 .apply(to: updatedObject)
             guard let originalData = data,
                   let original = try? ParseCoding.jsonDecoder().decode(T.self,
                                                                        from: originalData),
-                  original.hasSameObjectId(as: object) else {
-                      return object
+                  original.hasSameObjectId(as: updatedObject) else {
+                      return updatedObject
                   }
-            return try object.merge(with: original)
+            return try updatedObject.merge(with: original)
         }
         return API.Command<T, T>(method: .PATCH,
                                  path: object.endpoint,