+++ Sources/ParseSwift/Types/ParseOperation.swift
@@ -27,6 +27,53 @@ public struct ParseOperation<T>: Savable where T: ParseObject {
     }
 
     /**
+     An operation that sets a field's value.
+     - Parameters:
+        - keyPath: The respective `KeyPath` of the object.
+        - value: The value to set the `KeyPath` to.
+        - returns: The updated operations.
+     - warning: Do not combine operations using this method with other operations that
+     do not use this method to **set** all operations. If you need to combine multiple types
+     of operations such as: add, increment, forceSet, etc., use
+     `func set<W>(_ key: (String, WritableKeyPath<T, W?>), value: W?)`
+     instead.
+     */
+    @available(*, deprecated, message: "replace \"value\" with \"to\"")
+    public func set<W>(_ keyPath: WritableKeyPath<T, W?>,
+                       value: W) throws -> Self where W: Encodable & Equatable {
+        try set(keyPath, to: value)
+    }
+
+    /**
+     An operation that sets a field's value.
+     - Parameters:
+        - keyPath: The respective `KeyPath` of the object.
+        - to: The value to set the `KeyPath` to.
+        - returns: The updated operations.
+     - warning: Do not combine operations using this method with other operations that
+     do not use this method to **set** all operations. If you need to combine multiple types
+     of operations such as: add, increment, forceSet, etc., use
+     `func set<W>(_ key: (String, WritableKeyPath<T, W?>), value: W?)`
+     instead.
+     */
+    public func set<W>(_ keyPath: WritableKeyPath<T, W?>,
+                       to value: W) throws -> Self where W: Encodable & Equatable {
+        guard operations.isEmpty,
+              keysToNull.isEmpty else {
+            throw ParseError(code: .unknownError,
+                             message: """
+                                Cannot combine other operations such as: add, increment,
+                                forceSet, etc., with this method. Use the \"set\" method that takes
+                                the (String, WritableKeyPath) tuple as an argument instead to
+                                combine multiple types of operations.
+                                """)
+        }
+        var mutableOperation = self
+        mutableOperation.target = mutableOperation.target.set(keyPath, to: value)
+        return mutableOperation
+    }
+
+    /**
      An operation that sets a field's value if it has changed from its previous value.
      - Parameters:
         - key: A tuple consisting of the key and the respective `KeyPath` of the object.
@@ -34,8 +81,22 @@ public struct ParseOperation<T>: Savable where T: ParseObject {
         - returns: The updated operations.
      - Note: Set the value to "nil" if you want it to be "null" on the Parse Server.
      */
+    @available(*, deprecated, message: "replace \"value\" with \"to\"")
     public func set<W>(_ key: (String, WritableKeyPath<T, W?>),
                        value: W?) -> Self where W: Encodable & Equatable {
+        set(key, to: value)
+    }
+
+    /**
+     An operation that sets a field's value if it has changed from its previous value.
+     - Parameters:
+        - key: A tuple consisting of the key and the respective `KeyPath` of the object.
+        - to: The value to set the `KeyPath` to.
+        - returns: The updated operations.
+     - Note: Set the value to "nil" if you want it to be "null" on the Parse Server.
+     */
+    public func set<W>(_ key: (String, WritableKeyPath<T, W?>),
+                       to value: W?) -> Self where W: Encodable & Equatable {
         var mutableOperation = self
         if value == nil && target[keyPath: key.1] != nil {
             mutableOperation.keysToNull.insert(key.0)
@@ -57,6 +118,19 @@ public struct ParseOperation<T>: Savable where T: ParseObject {
      */
     public func forceSet<W>(_ key: (String, WritableKeyPath<T, W?>),
                             value: W?) -> Self where W: Encodable {
+        forceSet(key, to: value)
+    }
+
+    /**
+     An operation that force sets a field's value.
+     - Parameters:
+        - key: A tuple consisting of the key and the respective `KeyPath` of the object.
+        - to: The value to set the `KeyPath` to.
+        - returns: The updated operations.
+     - Note: Set the value to "nil" if you want it to be "null" on the Parse Server.
+     */
+    public func forceSet<W>(_ key: (String, WritableKeyPath<T, W?>),
+                            to value: W?) -> Self where W: Encodable {
         var mutableOperation = self
         if value != nil {
             mutableOperation.operations[key.0] = value