+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/4.10.0...main)
 * _Contributing to this repo? Add info about your change here to be included in the next release_
 
+### 4.11.0
+[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/4.10.0...4.11.0)
+
+__New features__
+- Add a set method that developers can call on their ParseObjects which automatically sends updated properties to a Parse Server and merges those updates with the original ParseObject locally. The feature removes the requirement to call mergeable and implement merge(), but comes at additional computational overhead ([#406](https://github.com/parse-community/Parse-Swift/pull/406)), thanks to [Corey Baker](https://github.com/cbaker6).
+
 ### 4.10.0
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/4.9.3...4.10.0)